refactor: Splited utils into multiple files
